backend_prep2pro/app/redis_client.py
import redis # type: ignore
from dotenv import load_dotenv # type: ignore
import os

# Load environment variables
load_dotenv()

# Set up the Redis client
REDIS_HOST = os.getenv('REDIS_HOST', 'localhost')
REDIS_PORT = os.getenv('REDIS_PORT', 6379)
REDIS_DB = os.getenv('REDIS_DB', 0)

# Create the Redis client
redis_client = redis.StrictRedis(host=REDIS_HOST, port=REDIS_PORT, db=REDIS_DB, decode_responses=True)