backend_prep2pro/requirements.txt
Flask
PyPDF2
Werkzeug
Flask-Caching
sentence_transformers
language_tool_python
google.generativeai
scikit-learn
python-dotenv
supabase
redis
langchain_google_genai
pandas
langchain_core
numpy
io
faiss-cpu
random